backend/requirements/base.txt
django==5.0.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
django-environ==0.11.2
django-extensions==3.2.3
django-filter==24.2
celery==5.3.6
redis==5.0.4
django-celery-beat==2.6.0
django-celery-results==2.5.1
GDAL==3.8.4
Pillow==10.3.0
django-storages==1.14.3
boto3==1.34.84
cloudinary==1.40.0
django-cloudinary-storage==0.3.0
djangorestframework-gis==1.0
requests==2.31.0
python-dateutil==2.9.0
